+++ kinto/plugins/quotas/__init__.py
@@ -11,7 +11,7 @@ def includeme(config):
 
     # If StatsD is enabled, monitor execution time of listener.
     if config.registry.statsd:
-        key = 'listeners.quotas'
+        key = 'plugins.quotas'
         listener = config.registry.statsd.timer(key)(on_resource_changed)
     else:
         listener = on_resource_changed